+++ tests/test_output/doc-test/test-root/conf.py
@@ -51,6 +51,7 @@ intersphinx_mapping = {"python": ("https://docs.python.org/3/", None)}
 
 rst_prolog = ".. |hello| replace:: world"
 
+
 # These revert https://github.com/sphinx-doc/sphinx/pull/8472
 def visit_desc_signature(self, node: addnodes.desc_signature) -> None:
 	# the id is set automatically